Move doExport out to a test utility, test visibility of loading spinner
